+++ transform.v
@@ -20,24 +20,24 @@ pub fn (val RGB) invert() RGB {
 }
 
 // lighten changes lightness of a given HSL struct and returns new struct
-pub fn (val HSL) lighten(value f32) HSL {
+pub fn (val HSL) lighten(value f64) HSL {
 	l := between(val.l + value, 0, 1)
 	return HSL { val.h, val.s, l }
 }
 
 // darken changes lightness of a given HSL struct and returns new struct
-pub fn (val HSL) darken(value f32) HSL {
+pub fn (val HSL) darken(value f64) HSL {
 	return val.lighten(-value)
 }
 
 // saturate changes saturation of a given HSL struct and returns new struct
-pub fn (val HSL) saturate(value f32) HSL {
+pub fn (val HSL) saturate(value f64) HSL {
 	s := between(val.s + value, 0, 1)
 	return HSL { val.h, s, val.l }
 }
 
 // desaturate changes saturation of a given HSL struct and returns new struct
-pub fn (val HSL) desaturate(value f32) HSL {
+pub fn (val HSL) desaturate(value f64) HSL {
 	return val.saturate(-value)
 }
